Rename SpotifyConnection.java Begin separating requests into individual classes
